enhanced the example...

+++ user_space/filesystem/sparse.cc
@@ -7,15 +7,12 @@
  * - other pages are full of 0's.
  *
  * 	Mark Veltzer
- *
- * TODO:
- * - show that reading from the file from the begining is possible and produces 0's.
  */
 
 #include <stdio.h>
 #include <unistd.h>
 #include <stdlib.h>
-#include <string.h>
+#include <string.h> // for memset(3)
 #include <sys/types.h> // for stat(2)
 #include <sys/stat.h> // for stat(2)
 #include <unistd.h> // for stat(2)
@@ -34,7 +31,6 @@ int main(int argc, char **argv, char **envp) {
 	// this will be used to check return values
 	int ret = 0;
 	FILE *f = fopen(fname, "w");
-
 	if (f == NULL) {
 		perror("could not open file");
 		exit(1);
@@ -45,8 +41,8 @@ int main(int argc, char **argv, char **envp) {
 		exit(1);
 	}
 	ret = fwrite(string, strlen(string), 1, f);
-	if (ret == -1) {
-		perror("could not read file");
+	if (ret != 1) {
+		perror("could not write file");
 		exit(1);
 	}
 	ret = fclose(f);
@@ -54,13 +50,33 @@ int main(int argc, char **argv, char **envp) {
 		perror("could not close file");
 		exit(1);
 	}
-	struct stat buf;
-	scie(stat(fname, &buf),"stat");
-	fprintf(stderr,"buf.size is %lu\n",buf.st_size);
-	fprintf(stderr,"buf.blocks (512B each) is %lu\n",buf.st_blocks);
-	fprintf(stderr,"real total calculated size is %lu\n",buf.st_blocks*512);
+	struct stat stat_buf;
+	scie(stat(fname, &stat_buf),"stat");
+	fprintf(stderr,"buf.size is %lu\n",stat_buf.st_size);
+	fprintf(stderr,"buf.blocks (512B each) is %lu\n",stat_buf.st_blocks);
+	fprintf(stderr,"real total calculated size is %lu\n",stat_buf.st_blocks*512);
 	my_system("ls -l %s",fname);
 	my_system("du -h %s",fname);
+	// now lets see that we can read from the file...
+	f = fopen(fname, "r");
+	if (f == NULL) {
+		perror("could not open file");
+		exit(1);
+	}
+	const unsigned int buf_size=4096;
+	char buf[buf_size];
+	memset(buf,3,buf_size);
+	memcheck(buf,3,buf_size);
+
+	ret = fread(buf, buf_size, 1, f);
+	if (ret != 1) {
+		perror("could not read file");
+		exit(1);
+	}
+	fprintf(stderr,"checking that the buffer is full of 0's\n");
+	memcheck(buf,0,buf_size);
+
+	// now lets get ridd of the file...
 	scie(unlink(fname),"unlink");
 	return(0);
 }
